Update st_foscam_hd.groovy
Edited the alarm off and alarm off definitions such that they do not lose the other configurations when turning the alarm off and on again. 

This is gone with a get command to get the entire configuration and then setting (or resetting) the isEnable bit, then writing back the entire configuration.